services/gateway/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
filelock==3.20.0
fsspec==2025.9.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
inquirerpy==0.3.4
Jinja2==3.1.6
jwt==1.4.0
Mako==1.3.10
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
passlib==1.7.4
pfzy==0.3.4
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg==3.3.3
psycopg-binary==3.3.3
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
sympy==1.13.1
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
uvicorn==0.41.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
